Seperate the classes into different files
