+++ app/components/home/homeView.js
@@ -1,4 +1,7 @@
 import React from 'react';
+import Tabs from './tabs';
+import Event from './event';
+import Collapsable from './collapsable';
 
 class HomeView extends React.Component {
   constructor(props) {
@@ -51,72 +54,11 @@ class HomeView extends React.Component {
       </div>
     );
   }
-
   render() {
     return (
       <div className="home-body">
         <Tabs />
         {this.createView.bind(this).call()}
-      </div>
-    );
-  }
-}
-class Tabs extends React.Component {
-  render() {
-    return (
-      <div className="event-tabs">
-        <ul className="nav nav-tabs justify-content-around">
-          <Tab name="Ongoing" />
-          <Tab name="Upcoming" />
-          <Tab name="Past" />
-        </ul>
-      </div>
-    );
-  }
-}
-
-class Tab extends React.Component {
-  render() {
-    return (
-      <div className="event-tab">
-        <li className="nav-item">
-          <a className="nav-link" href="#">{this.props.name}</a>
-        </li>
-      </div>
-    );
-  }
-}
-
-class Event extends React.Component {
-  handleClick(event) {
-    this.props.openCollapsable(this.props.serial);
-  }
-  render() {
-    return (
-      <div>
-        <div onClick={this.handleClick.bind(this)} className="event-picture" href={"#collapseExample"+this.props.serial}
-           aria-expanded={this.props.open[this.props.serial]} aria-controls={"collapseExample"+this.props.serial}>
-          <img className="img-fluid text-center event-poster" src="https://goo.gl/T2k8fq" alt="Event picture" />
-          <div id="event-poster" />
-        </div>
-    </div>
-    );
-  }
-}
-
-class Collapsable extends React.Component {
-  render() {
-    let showCollapse = (this.props.open[this.props.serial]) ? 'collapse show' : 'collapse';
-    return (
-      <div className={showCollapse} id={"collapseExample"+this.props.serial-1} aria-expanded={this.props.open[this.props.serial]}>
-        <div className="card card-block">
-          <div>
-            <h6>Title {this.props.serial}</h6>
-            <div id="event-title" />
-            <h6>Description {this.props.serial}</h6>
-            <div id="event-description" />
-          </div>
-        </div>
       </div>
     );
   }
